LPS-78225 Move JournalActivityKeys to api
