+++ modules/apps/web-experience/journal/journal-web/src/main/java/com/liferay/journal/web/social/JournalArticleActivityInterpreter.java
@@ -14,12 +14,12 @@
 
 package com.liferay.journal.web.social;
 
+import com.liferay.journal.constants.JournalActivityKeys;
 import com.liferay.journal.constants.JournalPortletKeys;
 import com.liferay.journal.model.JournalArticle;
 import com.liferay.journal.model.JournalArticleConstants;
 import com.liferay.journal.model.JournalFolder;
 import com.liferay.journal.service.JournalArticleLocalService;
-import com.liferay.journal.social.JournalActivityKeys;
 import com.liferay.journal.web.util.JournalResourceBundleLoader;
 import com.liferay.portal.kernel.model.Layout;
 import com.liferay.portal.kernel.security.permission.ActionKeys;
